+++ qm-music-parent/qm-music-core/src/main/java/com/github/chenqimiao/request/StarOrNotRequest.java
@@ -15,11 +15,11 @@ import lombok.*;
 @NoArgsConstructor
 public class StarOrNotRequest {
 
-    private Integer relationId;
+    private Long relationId;
 
     private EnumUserStarType startType;
 
-    private Integer userId;
+    private Long userId;
 
     /**
      * @see  com.github.chenqimiao.enums.EnumStarActionType
